+++ lib/db/src/schema/index.ts
@@ -20,3 +20,4 @@
 export * from "./initiatives";
 export * from "./validations";
 export * from "./calculations";
+export * from "./providerTests";